Task4.Double replaced by BigDecimal in ProductEntity.

+++ src/main/java/ru/ldwx/crm/model/ProductEntity.java
@@ -1,5 +1,6 @@
 package ru.ldwx.crm.model;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 
 public class ProductEntity {
@@ -8,10 +9,10 @@ public class ProductEntity {
     private String productNumber;
     private String productName;
     private String supplierLink;
-    private Double price;
+    private BigDecimal price;
     private Integer quantity;
 
-    public ProductEntity(Integer id, String productNumber, String productName, String supplierLink, Double price, Integer quantity) {
+    public ProductEntity(Integer id, String productNumber, String productName, String supplierLink, BigDecimal price, Integer quantity) {
         this.id = id;
         this.productNumber = productNumber;
         this.productName = productName;
@@ -20,7 +21,7 @@ public class ProductEntity {
         this.quantity = quantity;
     }
 
-    public ProductEntity(String productName, String supplierLink, Double price) {
+    public ProductEntity(String productName, String supplierLink, BigDecimal price) {
         this.productName = productName;
         this.supplierLink = supplierLink;
         this.price = price;
@@ -58,11 +59,11 @@ public class ProductEntity {
         this.supplierLink = supplierLink;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
